requirements.txt
asgiref==3.4.1
auth==0.5.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
eventlet==0.32.0
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
mongoengine==0.23.1
pycparser==2.20
pymongo==3.12.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==2.0.1
